restructured the file and added new file for new component
