+++ src/component/BodyComponent.jsx
@@ -1,10 +1,10 @@
 import RestaurantCard from "./RestaurantCard";
-import restaurantData from "./api/exampleData";
+import restaurantData from "../utils/mockData";
 
 const BodyComponent = () => {
   return (
     <div id="body-container" className="body">
-      <div className="search">search</div>
+      <div className="search">Search</div>
       <div className="restaurant-container">
         {restaurantData.map((resData) => (
           <RestaurantCard key={resData.info.id} resData={resData} />
