+++ src/component/Footer.jsx
@@ -1,0 +1,9 @@
+const FooterComponent = () => {
+  return (
+    <div className="footer">
+      &copy; 2023 Food Ordering App. All rights reserved.
+    </div>
+  );
+};
+
+export default FooterComponent;
