+++ src/component/RestaurantCard.jsx
@@ -1,3 +1,5 @@
+import { CDN_URL } from "../utils/constants";
+
 const RestaurantCard = ({ resData }) => {
   // const { resName, cuisine } = props;
   const {
@@ -12,11 +14,11 @@ const RestaurantCard = ({ resData }) => {
   } = resData?.info;
 
   return (
-    <div className="restaurant-card">
+    <div id={id} className="restaurant-card">
       <img
         alt="rest-image"
         className="res-logo"
-        src={`https://media-assets.swiggy.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_660/${cloudinaryImageId}`}
+        src={`${CDN_URL + cloudinaryImageId}`}
       />
       <h3>{name}</h3>
       <h4>{cuisines.join(", ")}</h4>
